refactor: move ActivityIndicator to utils.py
